Share Reciprocal across Uint and BoxedUint
