+++ src/lib.rs
@@ -181,7 +181,7 @@ pub use crate::{
     limb::{Limb, WideWord, Word},
     non_zero::NonZero,
     traits::*,
-    uint::div_limb::Reciprocal,
+    uint::reciprocal::Reciprocal,
     uint::*,
     wrapping::Wrapping,
 };
